Refactor code for clarity
Code had become one long block. This commit cleans that up and separates it following the same pattern as aiosmtpd for clarity.
