+++ bebleo_smtpd_fixture/__init__.py
@@ -1,2 +1,6 @@
 # Flake8: noqa
 from .fixture import smtpd
+from .controller import AuthController
+from .authenticator import Authenticator
+from .smtp import AuthSMTP
+from .handlers import AuthMessage
